get_result_point to backend, plotting import
