+++ src/plotting.py
@@ -10,7 +10,6 @@ import copy
 
 from beat import utility
 from beat.models import Stage
-from beat.sampler.metropolis import get_trace_stats
 from beat.heart import init_seismic_targets, init_geodetic_targets
 from beat.colormap import slip_colormap
 
@@ -472,60 +471,6 @@ def plot_log_cov(cov_mat):
     im = ax.imshow(num.multiply(num.log(num.abs(cov_mat)), mask))
     plt.colorbar(im)
     plt.show()
-
-
-def get_result_point(stage, config, point_llk='max'):
-    """
-    Return point of a given stage result.
-
-    Parameters
-    ----------
-    stage : :class:`models.Stage`
-    config : :class:`config.BEATConfig`
-    point_llk : str
-        with specified llk(max, mean, min).
-
-    Returns
-    -------
-    dict
-    """
-    if config.sampler_config.name == 'Metropolis':
-        if stage.step is None:
-            raise AttributeError(
-                'Loading Metropolis results requires'
-                ' sampler parameters to be loaded!')
-
-        sc = config.sampler_config.parameters
-        pdict, _ = get_trace_stats(
-            stage.mtrace, stage.step, sc.burn, sc.thin)
-        point = pdict[point_llk]
-
-    elif config.sampler_config.name == 'SMC':
-        llk = stage.mtrace.get_values(
-            varname='like',
-            combine=True)
-
-        posterior_idxs = utility.get_fit_indexes(llk)
-
-        point = stage.mtrace.point(idx=posterior_idxs[point_llk])
-
-    elif config.sampler_config.name == 'PT':
-        params = config.sampler_config.parameters
-        llk = stage.mtrace.get_values(
-            varname='like',
-            burn=int(params.n_samples * params.burn),
-            thin=params.thin)
-
-        posterior_idxs = utility.get_fit_indexes(llk)
-
-        point = stage.mtrace.point(idx=posterior_idxs[point_llk])
-
-    else:
-        raise NotImplementedError(
-            'Sampler "%s" is not supported!' % config.sampler_config.name)
-
-    return point
-
 
 def plot_quadtree(ax, data, target, cmap, colim, alpha=0.8):
     """
